requirements.txt
CppHeaderParser>=2.7.4
build>=1.2.2
meson>=1.7.0
pre-commit>=4.3.0
python-magic>=0.4.27; platform_system != "Windows"
PyYAML==6.0.2
pyzstd>=0.16.0
# `pkg_resources` has been removed from Setuptools in version 82.0.0
# but it still a dependency in some environments (issue#3311)
setuptools>=80.9.0,<82.0.0
tomli==2.2.1; python_version <= '3.10'
pytest-cmake==0.13.0
mako>=1.1.3
# Compatible with aiobotocore 2.26.0 from dvc[s3] 3.62.0
# installed in build_manylinux_x86_64.Dockerfile
boto3>=1.41.0,<1.42.0

# hipBLASLt deps
joblib>=1.4.2
msgpack>=1.1.0

# KPACK
zstandard>=0.19.0

# libhipcxx build requirement
lit==18.1.8
